Fix template.cpp

+++ template.cpp
@@ -1,13 +1,13 @@
 #include <iostream>
 
 #define rep(i, n) for (int i = 0; i < (int)(n); i++)
-#define int long long
+#define ll long long
 const long long INF = 1LL << 60;
 
 using namespace std;
 
 signed main() {
-  int N;
+  ll N;
   cin >> N;
 
   return 0;
